The big reformatting commit...

+++ src/NodaTime/Fields/BasicMonthOfYearDateTimeField.cs
@@ -14,6 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 #endregion
+
 using System;
 using NodaTime.Calendars;
 using NodaTime.Partials;
@@ -44,7 +45,6 @@ namespace NodaTime.Fields
         public override bool IsLenient { get { return false; } }
 
         #region Values
-
         /// <summary>
         /// Get the Month component of the specified local instant.
         /// </summary>
@@ -196,7 +196,6 @@ namespace NodaTime.Fields
             // Get proper date part, and return result
             long datePart = calendarSystem.GetYearMonthDayTicks(intYearToUse, intMonthToUse, dayToUse);
             return new LocalInstant(datePart + timePart);
-
         }
 
         /// <summary>
@@ -285,25 +284,21 @@ namespace NodaTime.Fields
             int month = (int)value;
             int thisYear = calendarSystem.GetYear(localInstant);
             int thisDom = calendarSystem.GetDayOfMonth(localInstant, thisYear);
-            int maxDom = calendarSystem.GetDaysInYearMonth(thisYear, (int)month);
+            int maxDom = calendarSystem.GetDaysInYearMonth(thisYear, month);
             if (thisDom > maxDom)
             {
                 // Quietly force DOM to nearest sane value.
                 thisDom = maxDom;
             }
-            return new LocalInstant(calendarSystem.GetYearMonthDayTicks(thisYear, month, thisDom) +
-                calendarSystem.GetTickOfDay(localInstant));
+            return new LocalInstant(calendarSystem.GetYearMonthDayTicks(thisYear, month, thisDom) + calendarSystem.GetTickOfDay(localInstant));
         }
-
         #endregion
 
         #region Leap
-
         public override bool IsLeap(LocalInstant localInstant)
         {
             int thisYear = calendarSystem.GetYear(localInstant);
-            return calendarSystem.IsLeapYear(thisYear) &&
-                calendarSystem.GetMonthOfYear(localInstant, thisYear) == leapMonth;
+            return calendarSystem.IsLeapYear(thisYear) && calendarSystem.GetMonthOfYear(localInstant, thisYear) == leapMonth;
         }
 
         public override int GetLeapAmount(LocalInstant localInstant)
@@ -312,11 +307,9 @@ namespace NodaTime.Fields
         }
 
         public override IDurationField LeapDurationField { get { return calendarSystem.Fields.Days; } }
-
         #endregion
 
         #region Ranges
-
         public override long GetMinimumValue()
         {
             return MinimumValue;
@@ -326,11 +319,9 @@ namespace NodaTime.Fields
         {
             return max;
         }
-
         #endregion
 
         #region Rounding
-
         public override LocalInstant RoundFloor(LocalInstant localInstant)
         {
             int year = calendarSystem.GetYear(localInstant);
@@ -342,7 +333,6 @@ namespace NodaTime.Fields
         {
             return localInstant - RoundFloor(localInstant);
         }
-
         #endregion
     }
 }
